# backend/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Life Circle Backend",
    "version": "2.0.0"
  },
  "paths": {
    "/api/v1/analysis/hybrid": {
      "post": {
        "tags": [
          "Hybrid Isochrone"
        ],
        "summary": "Create",
        "operationId": "create_api_v1_analysis_hybrid_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HybridRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskStatusResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HybridError"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HybridError"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HybridError"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HybridError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/analyses": {
      "post": {
        "tags": [
          "Hybrid Isochrone"
        ],
        "summary": "Create",
        "operationId": "create_api_analyses_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HybridRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskStatusResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HybridError"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HybridError"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HybridError"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HybridError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HybridConfig": {
        "properties": {
          "time_limit_seconds": {
            "type": "integer",
            "const": 900,
            "title": "Time Limit Seconds",
            "default": 900
          },
          "endpoint_offset_limit_m": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Endpoint Offset Limit M",
            "default": 50
          },
          "initial_direction_count": {
            "type": "integer",
            "maximum": 64.0,
            "minimum": 4.0,
            "title": "Initial Direction Count",
            "default": 16
          },
          "analysis_half_width_m": {
            "type": "number",
            "maximum": 1200.0,
            "exclusiveMinimum": 0.0,
            "title": "Analysis Half Width M",
            "default": 1200
          },
          "initial_probe_distances_m": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "title": "Initial Probe Distances M",
            "default": [
              400,
              700,
              1000,
              1200
            ]
          },
          "max_exploration_radius_m": {
            "type": "number",
            "maximum": 10000.0,
            "exclusiveMinimum": 0.0,
            "title": "Max Exploration Radius M",
            "default": 3200
          },
          "boundary_tolerance_m": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Boundary Tolerance M",
            "default": 50
          },
          "max_refinement_depth": {
            "type": "integer",
            "maximum": 12.0,
            "minimum": 0.0,
            "title": "Max Refinement Depth",
            "default": 6
          },
          "max_baidu_requests": {
            "type": "integer",
            "maximum": 400.0,
            "minimum": 1.0,
            "title": "Max Baidu Requests",
            "default": 400
          },
          "max_direction_count": {
            "type": "integer",
            "maximum": 256.0,
            "minimum": 4.0,
            "title": "Max Direction Count",
            "default": 64
          },
          "angular_refinement_threshold_m": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Angular Refinement Threshold M",
            "default": 250
          },
          "angular_refinement_ratio": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Angular Refinement Ratio",
            "default": 0.25
          },
          "topology_risk_threshold": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Topology Risk Threshold",
            "default": 0.6
          },
          "topology_search_radius_m": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Topology Search Radius M",
            "default": 150
          },
          "max_triangle_edge_m": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Max Triangle Edge M",
            "default": 400
          },
          "mixed_triangle_target_m": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Mixed Triangle Target M",
            "default": 100
          },
          "request_qps": {
            "type": "number",
            "maximum": 20.0,
            "exclusiveMinimum": 0.0,
            "title": "Request Qps",
            "default": 3
          },
          "deadline_seconds": {
            "type": "number",
            "maximum": 7200.0,
            "exclusiveMinimum": 0.0,
            "title": "Deadline Seconds",
            "default": 1800
          },
          "failure_streak_limit": {
            "type": "integer",
            "maximum": 20.0,
            "minimum": 1.0,
            "title": "Failure Streak Limit",
            "default": 5
          },
          "origin_endpoint_failure_streak_limit": {
            "type": "integer",
            "maximum": 10.0,
            "minimum": 1.0,
            "title": "Origin Endpoint Failure Streak Limit",
            "default": 3
          },
          "interior_gap_radii_m": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "title": "Interior Gap Radii M",
            "default": [
              25,
              50,
              100
            ]
          },
          "bridge_display_width_m": {
            "type": "number",
            "maximum": 10.0,
            "exclusiveMinimum": 0.0,
            "title": "Bridge Display Width M",
            "default": 3
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "HybridConfig"
      },
      "HybridError": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "HybridError"
      },
...
      "HybridRequest": {
        "properties": {
          "origin": {
            "$ref": "#/components/schemas/Origin"
          },
          "coordinate_system": {
            "type": "string",
            "const": "bd09ll",
            "title": "Coordinate System"
          },
          "config": {
            "$ref": "#/components/schemas/HybridConfig"
          },
          "client_request_id": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "pattern": "^[a-zA-Z0-9_-]+$",
            "title": "Client Request Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "origin",
          "coordinate_system",
          "client_request_id"
        ],
        "title": "HybridRequest"
      },
...
      "Origin": {
        "properties": {
          "lng": {
            "type": "number",
            "maximum": 180.0,
            "minimum": -180.0,
            "title": "Lng"
          },
          "lat": {
            "type": "number",
            "exclusiveMaximum": 85.0,
            "exclusiveMinimum": -85.0,
            "title": "Lat"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "lng",
          "lat"
        ],
        "title": "Origin"
      },
...
      "TaskStatusResponse": {
        "properties": {
          "schema_version": {
            "type": "string",
            "const": "1.0",
            "title": "Schema Version",
            "default": "1.0"
          },
          "responseType": {
            "type": "string",
            "const": "task",
            "title": "Responsetype",
            "default": "task"
          },
          "taskId": {
            "type": "string",
            "title": "Taskid"
          },
          "status": {
            "type": "string",
            "enum": [
              "running",
              "cancelling",
              "completed",
              "cancelled",
              "failed"
            ],
            "title": "Status"
          },
          "businessStatus": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "complete",
                  "partial",
                  "failed",
                  "empty"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Businessstatus"
          },
          "stage": {
            "type": "string",
            "title": "Stage"
          },
          "requests": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Requests"
          },
          "networkRequests": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Networkrequests"
          },
          "budget": {
            "type": "integer",
            "maximum": 800.0,
            "minimum": 1.0,
            "title": "Budget"
          },
          "elapsedSeconds": {
            "type": "number",
            "minimum": 0.0,
            "title": "Elapsedseconds"
          },
          "dataSource": {
            "type": "string",
            "enum": [
              "synthetic",
              "baidu_walking"
            ],
            "title": "Datasource"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "taskId",
          "status",
          "stage",
          "requests",
          "networkRequests",
          "budget",
          "elapsedSeconds",
          "dataSource"
        ],
        "title": "TaskStatusResponse",
        "description": "统一异步任务响应；business_status 是业务结果，status 是传输任务状态。"
      }
    }
  }
}